+++ src/scriptlets/scriptlets-list.ts
@@ -76,6 +76,7 @@ export { trustedReplaceOutboundText } from './trusted-replace-outbound-text';
 export { preventCanvas } from './prevent-canvas';
 export { trustedReplaceArgument } from './trusted-replace-argument';
 export { preventInnerHTML } from './prevent-innerHTML';
+export { preventConstructor } from './prevent-constructor';
 // redirects as scriptlets
 // https://github.com/AdguardTeam/Scriptlets/issues/300
 export { AmazonApstag } from './amazon-apstag';