+++ src/scriptlets/scriptlets-names-list.ts
@@ -76,6 +76,7 @@ export { trustedReplaceOutboundTextNames } from './trusted-replace-outbound-text
 export { preventCanvasNames } from './prevent-canvas';
 export { trustedReplaceArgumentNames } from './trusted-replace-argument';
 export { preventInnerHTMLNames } from './prevent-innerHTML';
+export { preventConstructorNames } from './prevent-constructor';
 // redirects as scriptlets
 // https://github.com/AdguardTeam/Scriptlets/issues/300
 export { AmazonApstagNames } from './amazon-apstag';